+++ pyam/plotting.py
@@ -384,7 +384,7 @@ def pie_plot(df, value='value', category='variable',
 
 
 def stack_plot(df, x='year', y='value', stack='variable',
-               ax=None, legend=True, title=True, cmap=None,
+               ax=None, legend=True, title=True, cmap=None, total=None,
                **kwargs):
     """Plot data as a stack chart.
 
@@ -410,6 +410,11 @@ def stack_plot(df, x='year', y='value', stack='variable',
     cmap : string, optional
         A colormap to use.
         default: None
+    total : bool or dict, optional
+        If True, plot a total line with default pyam settings. If a dict, then
+        plot the total line using the dict key-value pairs as keyword arguments
+        to ax.plot(). If None, do not plot the total line.
+        default : None
     kwargs : Additional arguments to pass to the pd.DataFrame.plot() function
     """
     for col in set(SORT_IDX) - set([x, stack]):
@@ -423,19 +428,99 @@ def stack_plot(df, x='year', y='value', stack='variable',
     # long form to one column per bar group
     _df = reshape_bar_plot(df, x, y, stack)
 
+    # Line below is for interpolation. On datetimes I think you'd downcast to
+    # seconds first and then cast back to datetime at the end..?
+    _df.index = _df.index.astype(float)
+
+    time_original = _df.index.values
+    first_zero_times = pd.DataFrame(index=["first_zero_time"])
+
+    both_positive_and_negative = _df.apply(
+        lambda x: (x >= 0).any() and (x < 0).any()
+    )
+    for col in _df.loc[:, both_positive_and_negative]:
+        values = _df[col].dropna().values
+        positive = (values >= 0)
+        negative = (values < 0)
+        pos_to_neg = positive[:-1] & negative[1:]
+        neg_to_pos = positive[1:] & negative[:-1]
+        crosses = np.argwhere(pos_to_neg | neg_to_pos)
+        for i, cross in enumerate(crosses):
+            cross = cross[0]  # get location
+            x_1 = time_original[cross]
+            x_2 = time_original[cross + 1]
+            y_1 = values[cross]
+            y_2 = values[cross + 1]
+
+            zero_time = x_1 - y_1 * (x_2 - x_1) / (y_2 - y_1)
+            if i == 0:
+                first_zero_times.loc[:, col] = zero_time
+            if zero_time not in _df.index.values:
+                _df.loc[zero_time, :] = np.nan
+
+    first_zero_times = first_zero_times.sort_values(
+        by="first_zero_time",
+        axis=1,
+    )
+    _df = _df.reindex(sorted(_df.index)).interpolate(method="values")
+
+    # Sort lines so that negative timeseries are on the right, positive
+    # timeseries are on the left and timeseries which go from positive to
+    # negative are ordered such that the timeseries which goes negative first
+    # is on the right (case of timeseries which go from negative to positive
+    # is an edge case we haven't thought about as it's unlikely to apply to
+    # us).
+    pos_cols = [c for c in _df if (_df[c] > 0).all()]
+    cross_cols = first_zero_times.columns[::-1].tolist()
+    neg_cols = [c for c in _df if (_df[c] < 0).all()]
+    col_order = pos_cols + cross_cols + neg_cols
+    _df = _df[col_order]
+
     # explicitly get colors
     defaults = default_props(reset=True, num_colors=len(_df.columns),
                              colormap=cmap)['color']
     rc = run_control()
-    colors = []
+    colors = {}
     for key in _df.columns:
         c = next(defaults)
-        if 'color' in rc and stack in rc['color'] and\
-                key in rc['color'][stack]:
+        c_in_rc = 'color' in rc
+        if c_in_rc and stack in rc['color'] and key in rc['color'][stack]:
             c = rc['color'][stack][key]
-        colors.append(c)
+        colors[key] = c
 
-    ax.stackplot(_df.index, _df.T, labels=_df.columns, colors=colors, **kwargs)
+    # plot stacks, starting from the top and working our way down to the bottom
+    negative_only_cumulative = _df.applymap(
+        lambda x: x if x < 0 else 0
+    ).cumsum(axis=1)
+    positive_only_cumulative = _df.applymap(lambda x: x if x >= 0 else 0)[
+        col_order[::-1]
+    ].cumsum(axis=1)[
+        col_order
+    ]
+    time = _df.index.values
+    upper = positive_only_cumulative.iloc[:, 0].values
+    for j, col in enumerate(_df):
+        noc_tr = negative_only_cumulative.iloc[:, j].values
+        try:
+            poc_nr = positive_only_cumulative.iloc[:, j + 1].values
+        except IndexError:
+            poc_nr = np.zeros_like(upper)
+        lower = poc_nr.copy()
+        if (noc_tr < 0).any():
+            lower[np.where(poc_nr == 0)] = noc_tr[np.where(poc_nr == 0)]
+
+        ax.fill_between(time, lower, upper, label=col,
+                        color=colors[col], **kwargs)
+        upper = lower.copy()
+
+    # add total
+    if (total is not None) and total:  # cover case where total=False
+        if isinstance(total, bool):  # can now assume total=True
+            total = {}
+        total.setdefault("label", "Total")
+        total.setdefault("color", "black")
+        total.setdefault("lw", 4.0)
+        ax.plot(time, _df.sum(axis=1), **total)
 
     # add legend
     ax.legend(loc='center left', bbox_to_anchor=(1.0, 0.5))